PWM LED fading.
LED was connected to PA5 to F411RE. Used tim1ch1 to generate pwm signal
to control the brightness. Set prescaler to 83 giving 1MHz, so 1 tick =
1 microsecond. Set ARR to 999 so that 1Mhz/(999+1) = 1000Hz, so PMW
period = 1ms and LED brightness updates every 1ms. CCR goes from 0 to
ARR (so 0-999), and duty cycle = CCR/(ARR+1) * 100%. So if CCR is set to
500 -> 50% duty cycle -> LED is only on half brightness.